+++ scdl/src/main/java/net/rdrei/android/scdl2/ui/DownloadActivity.java
@@ -3,24 +3,15 @@ package net.rdrei.android.scdl2.ui;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.Handler;
 import android.widget.Toast;
 
-import com.google.inject.Inject;
-
 import net.rdrei.android.scdl2.R;
-import net.rdrei.android.scdl2.api.APIException;
 import net.rdrei.android.scdl2.api.MediaDownloadType;
 import net.rdrei.android.scdl2.api.MediaState;
 import net.rdrei.android.scdl2.api.PendingDownload;
-import net.rdrei.android.scdl2.api.ServiceManager;
-import net.rdrei.android.scdl2.api.entity.TrackEntity;
-import net.rdrei.android.scdl2.api.service.PlaylistService;
-import net.rdrei.android.scdl2.api.service.TrackService;
 
 import roboguice.activity.RoboFragmentActivity;
 import roboguice.util.Ln;
-import roboguice.util.RoboAsyncTask;
 
 /**
  * Dispatches a download request to either the single download or playlist download activity.
@@ -83,24 +74,19 @@ public class DownloadActivity extends RoboFragmentActivity {
 
 		@Override
 		protected void onSuccess(PendingDownload download) throws Exception {
-			final MediaStateLoaderTask task = new MediaStateLoaderTask(getContext(), download);
+			final AbstractMediaStateLoaderTask task = new MediaStateLoaderTask(getContext(), download);
 			task.execute();
 		}
 	}
 
-	private class MediaStateLoaderTask extends RoboAsyncTask<MediaState> {
-		private final PendingDownload mPendingDownload;
-
-		@Inject
-		private ServiceManager mServiceManager;
-
-		@Inject
-		private roboguice.inject.ContextScope mContextScope;
+	/**
+	 * Takes a pending download and resolves it into a MediaState
+	 */
+	private class MediaStateLoaderTask extends AbstractMediaStateLoaderTask {
 
 		protected MediaStateLoaderTask(final Context context, final PendingDownload download) {
-			super(context);
+			super(context, download);
 			assert download.getType() == MediaDownloadType.TRACK;
-			mPendingDownload = download;
 		}
 
 		@Override
@@ -119,29 +105,5 @@ public class DownloadActivity extends RoboFragmentActivity {
 			loadMediaFragments();
 		}
 
-		@Override
-		public MediaState call() throws Exception {
-			mContextScope.enter(context);
-
-			try {
-				return resolveDownloadToMedia();
-			} finally {
-				mContextScope.exit(context);
-			}
-		}
-
-		private MediaState resolveDownloadToMedia() throws APIException {
-			switch (mPendingDownload.getType()) {
-				case PLAYLIST:
-					final PlaylistService playlistService = mServiceManager.playlistService();
-					return MediaState.fromEntity(
-							playlistService.getPlaylist(mPendingDownload.getId()));
-				case TRACK:
-					final TrackService trackService = mServiceManager.trackService();
-					return MediaState.fromEntity(trackService.getTrack(mPendingDownload.getId()));
-				default:
-					throw new IllegalStateException("Unknown PendingDownload type. WTF?");
-			}
-		}
 	}
 }
